chore: refactor login / create-account pages
