+++ src/types/profile.ts
@@ -1,0 +1,5 @@
+export type Profile = {
+	username: string
+	address: string
+	encryptedWallet: string
+}
